runner_agent/pyproject.toml
[project]
name = "lab2sc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles==23.2.1",
 "pillow",
 "spade==3.3.3",
]

[dependency-groups]
dev = ["ipdb>=0.13.13"]
